refactor(desktop): own element selector lifecycle
